update: 정수형 자료형 형변환으로 인한 잠재적 오버플로우 위협 제거

+++ src/main/kotlin/com/team/incube/gsmc/v3/domain/member/presentation/data/response/SearchMemberResponse.kt
@@ -4,6 +4,6 @@ import com.team.incube.gsmc.v3.domain.member.dto.Member
 
 data class SearchMemberResponse(
     val totalPage: Int,
-    val totalElements: Int,
+    val totalElements: Long,
     val data: List<Member>,
 )
